% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_delete_request}
\alias{api_delete_request}
\title{Perform a DELETE request to a URI on the Alveo server}
\usage{
api_delete_request(url)
}
\arguments{
\item{url}{The API URL that will be used for the request}
}
\value{
API response as json
}
\description{
Perform a DELETE request to a URI on the Alveo server
}
