% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_request}
\alias{api_request}
\title{api_request}
\usage{
api_request(url, data = NULL, binary = FALSE)
}
\arguments{
\item{url}{The API URL that will be used for the request}

\item{data}{Any data to be sent along with the request}

\item{binary}{A boolean, if TRUE then expect a binary result from the request}
}
\value{
API response as json
}
\description{
Perform a request for the given url, sending the API key along in the header, return the response
}
