% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Document-class}
\alias{Document}
\alias{Document-class}
\title{A class representing a single document from the Alveo virtual lab}
\description{
A class representing a single document from the Alveo virtual lab
}
\section{Fields}{

\describe{
\item{\code{uri}}{The URI of the document}

\item{\code{type}}{The document type (audio, video, text)}

\item{\code{size}}{The document size in bytes}
}}
\section{Methods}{

\describe{
\item{\code{download(destination = NULL)}}{Download the document either to the cache or a given destination directory, return the local name of the downloaded file}

\item{\code{get_content()}}{Get the document at the given URL, return value may be binary data}
}}

