% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_content}
\alias{document_content}
\title{Return the contents of a document}
\usage{
document_content(uri, binary = FALSE)
}
\arguments{
\item{uri}{The URI of the document on Alveo}

\item{binary}{If TRUE, download the data as binary, otherwise (default) download as text}
}
\value{
The content of the file
}
\description{
Return the contents of a document
}
