% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document_download}
\alias{document_download}
\title{Download a document to a given directory}
\usage{
document_download(uri, destination, binary = TRUE)
}
\arguments{
\item{uri}{The URI of the document on Alveo}

\item{destination}{The directory to store the downloaded file}

\item{binary}{If TRUE (default), download the data as binary, otherwise download as text}
}
\value{
The name of the local file
}
\description{
Download a document to a given directory
}
