% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterSegs}
\alias{filterSegs}
\title{Make a new segment list containing only the given labels}
\usage{
filterSegs(seglist, labels)
}
\arguments{
\item{seglist}{an Emu segment list}

\item{labels}{a vector of labels}
}
\value{
a new segment list containing only those labels
in the vector supplied
}
\description{
Make a new segment list containing only the given labels
}

