% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{Item-class}
\alias{Item}
\alias{Item-class}
\title{A class representing a single item from the Alveo virtual lab}
\description{
A class representing a single item from the Alveo virtual lab
}
\section{Fields}{

\describe{
\item{\code{id}}{The item identifier}

\item{\code{uri}}{The URI of the item}
}}
\section{Methods}{

\describe{
\item{\code{get_all_metadata()}}{Return full metadata for this item, including system properties}

\item{\code{get_annotations(type = NULL, label = NULL)}}{Return all annotations on this document, the 'type' and 'label' arguments can be used to restrict the annotations returned.}

\item{\code{get_document(index)}}{Return the document given by the (numerical) index, returns a Document object}

\item{\code{get_documents()}}{Return a list of all documents for this item}

\item{\code{get_indexable_text()}}{Return the indexable text for this item if any}

\item{\code{get_metadata()}}{Return the descriptive metadata for this item}
}}

