% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ItemList-class}
\alias{ItemList}
\alias{ItemList-class}
\title{A class representing an Alveo Item List}
\description{
A class representing an Alveo Item List
}
\section{Fields}{

\describe{
\item{\code{name}}{The item list name}

\item{\code{uri}}{The uri for the item list on the Alveo server}

\item{\code{items}}{A list of items in the item list}
}}
\section{Methods}{

\describe{
\item{\code{download(destination, format = "zip")}}{download all items in this item list, destination is the name of a directory to write the result in, format (zip or WARC or json). Returns the filename that is created.}

\item{\code{get_item(index)}}{Get an item given the (numerical) index, returns an Item object}

\item{\code{get_item_documents()}}{Return a vector of all of the documents for all of the items in this item list}

\item{\code{get_segment_list(type = "", label = "")}}{Query annotations on this item list and return an Emu segment list.  By default returns all annotations, specify the 'type' argument to restrict the annotation types, specify the 'label' argument to match specific labels}

\item{\code{num_items()}}{Return the number of items in this item list}
}}

