% Generated by roxygen2 (4.0.1): do not edit by hand
\name{localiseSegs}
\alias{localiseSegs}
\title{Download all files referenced by a segment list}
\usage{
localiseSegs(seglist)
}
\arguments{
\item{seglist}{an Emu segment list}
}
\value{
a new segment list that references the downloaded files
}
\description{
Take a segment list containing URLs, download the files and
replace the filenames with the local names, return a new
segment list
}

