% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_config}
\alias{read_config}
\title{read_config}
\usage{
read_config(config_file = NULL)
}
\arguments{
\item{config_file}{Location of the configuration file to read, default is 'alveo.config' in the home directory}
}
\description{
Read the user configuration file by default ~/alveo.config and return a hash of keys and values
}

