% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeItemFromCache}
\alias{removeItemFromCache}
\title{removeItemFromCache}
\usage{
removeItemFromCache(filename)
}
\arguments{
\item{filename}{the name of the local file to remove from the cache}
}
\description{
Removes requested file from the cache if it exists
}

