% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeCacheListing}
\alias{writeCacheListing}
\title{writeCacheListing}
\usage{
writeCacheListing(uri, filename)
}
\arguments{
\item{uri}{The URI being cached}

\item{filename}{The local filename that was used to cache the file}
}
\description{
Records the uri to filename mapping to the cache contents file
}

