% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{class}
\name{Document-class}
\alias{Document-class}
\alias{Document}
\title{A class representing a single document from the Alveo virtual lab}
\description{
A class representing a single document from the Alveo virtual lab
}
\section{Fields}{

\describe{
\item{\code{uri}}{The URI of the document}

\item{\code{type}}{The document type (audio, video, text)}

\item{\code{size}}{The document size in bytes}
}}

\section{Methods}{

\describe{
\item{\code{download(destination, binary = FALSE)}}{Download the document to a given destination directory, return the local name of the downloaded file}

\item{\code{get_content(binary = FALSE)}}{Get the document at the given URL, return value may be binary data}
}}

