% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\docType{class}
\name{Item-class}
\alias{Item-class}
\alias{Item}
\title{A class representing a single item from the Alveo virtual lab}
\description{
A class representing a single item from the Alveo virtual lab
}
\section{Fields}{

\describe{
\item{\code{id}}{The item identifier}

\item{\code{uri}}{The URI of the item}
}}

\section{Methods}{

\describe{
\item{\code{get_all_metadata()}}{Return full metadata for this item, including system properties}

\item{\code{get_annotations(type = NULL, label = NULL)}}{Return all annotations on this document, the 'type' and 'label' 
arguments can be used to restrict the annotations returned.}

\item{\code{get_document(index)}}{Return the document given by the (numerical) index, returns a Document object}

\item{\code{get_documents(types = NULL, pattern = NULL)}}{Return a list of all documents for this item
If type is given, it should be a sequence of type names, return only documents
of with dcterms:type in this sequence, eg. ('Audio', 'TextGrid').
If pattern is given, return only documents with dcterms:identifier matching this regular expression}

\item{\code{get_indexable_text()}}{Return the indexable text for this item if any}

\item{\code{get_metadata()}}{Return the descriptive metadata for this item}
}}

