% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_list.R
\docType{class}
\name{ItemList-class}
\alias{ItemList-class}
\alias{ItemList}
\title{A class representing an Alveo Item List}
\description{
A class representing an Alveo Item List
}
\section{Fields}{

\describe{
\item{\code{name}}{The item list name}

\item{\code{uri}}{The uri for the item list on the Alveo server}

\item{\code{items}}{A list of items in the item list}
}}

\section{Methods}{

\describe{
\item{\code{download(destination, type = NULL, pattern = NULL)}}{Download all items in this item list, destination is the name of a directory 
to write the result in, format (zip or WARC or json). 
Returns the filename that is created.}

\item{\code{get_item(index)}}{Get an item given the (numerical) index, returns an Item object}

\item{\code{get_item_documents(types = NULL, pattern = NULL)}}{Return a vector of all of the documents for all of the items in this item list.
If type is given, it should be a sequence of type names, return only documents
of with dc:type in this sequence, eg. ('Audio', 'TextGrid').
If pattern is given, return only documents with dc:identifier matching this regular expression}

\item{\code{num_items()}}{Return the number of items in this item list}
}}

